RECOMP-421: Automate Storybook deployment
There's probably a better way to do this with Try, but this should work.

The main issue is that we need to build Firefox to get
"browser/components/storybook.storybook/chrome-map.js". There may be
a way to have Try do this for us and then we can build the docs through
the GitHub action, but I haven't looked that far into that route.